Add .do file recoded variables
This is the .do file with all of the relevant GSS variables recoded to have a reference category at value 0 rather than the GSS's default value of 1. I also recoded the occupation variable to match the updated 2010 census occupational codes. Langdon and Klomegah were using outdated 1980 codes for their system. I still need to find a way to simplify this occupation variable because it has too many categories to interpret substantively very easily, but this is a good start.