New Occupation Var
I created a second version of the occupation variable (occ10) which took the 2010 SOC and made it look more like the 1980 SOC so as to reproduce the original study. The numbers look kind of wonky though and I don't think the new variable is substantively interesting because the original 1980 coding is frankly horrible. I think the original researchers may not have made the best decision. Will continue to review and revise as I find new ways to make this variable more rigorous.